feat: Add an abstraction for portal configs in PlayerSimulation3D
