+++ src/aux-server/aux-web/aux-player/scene/InventoryPortalConfig.ts
@@ -1,0 +1,84 @@
+import {
+    isDimensionLocked,
+    DEFAULT_PORTAL_ZOOMABLE,
+    DEFAULT_PORTAL_PANNABLE,
+    hasValue,
+    calculateBotValue,
+    BotCalculationContext,
+    PrecalculatedBot,
+    calculateGridScale,
+    calculateBooleanTagValue,
+    calculateNumericalTagValue,
+    DEFAULT_PORTAL_ROTATABLE,
+    DEFAULT_INVENTORY_PORTAL_RESIZABLE,
+    DEFAULT_INVENTORY_PORTAL_HEIGHT,
+} from '@casual-simulation/aux-common';
+import { Color } from 'three';
+import {
+    BrowserSimulation,
+    watchPortalConfigBot,
+} from '@casual-simulation/aux-vm-browser';
+import { tap } from 'rxjs/operators';
+import { SubscriptionLike, Subscription } from 'rxjs';
+import { PlayerGrid3D } from '../PlayerGrid3D';
+import { PortalConfig } from './PortalConfig';
+
+/**
+ * Defines a class that is able to watch dimension confic bots and update values.
+ */
+export class InventoryPortalConfig extends PortalConfig {
+    private _resizable: boolean;
+    private _height: number;
+
+    /**
+     * Gets whether the portal is resizable.
+     */
+    get resizable() {
+        if (this._resizable != null) {
+            return this._resizable;
+        } else {
+            return DEFAULT_INVENTORY_PORTAL_RESIZABLE;
+        }
+    }
+
+    /**
+     * Gets the height of the portal.
+     */
+    get height() {
+        if (this._height != null) {
+            return this._height;
+        } else {
+            return DEFAULT_INVENTORY_PORTAL_HEIGHT;
+        }
+    }
+
+    constructor(portalTag: string, simulation: BrowserSimulation) {
+        super(portalTag, simulation);
+    }
+
+    protected _clearPortalValues() {
+        super._clearPortalValues();
+        this._resizable = null;
+        this._height = null;
+    }
+
+    protected _updatePortalValues(
+        calc: BotCalculationContext,
+        bot: PrecalculatedBot,
+        portalTag: string
+    ) {
+        super._updatePortalValues(calc, bot, portalTag);
+        this._resizable = calculateBooleanTagValue(
+            calc,
+            bot,
+            `auxInventoryPortalResizable`,
+            DEFAULT_INVENTORY_PORTAL_RESIZABLE
+        );
+        this._height = calculateNumericalTagValue(
+            calc,
+            bot,
+            `auxInventoryPortalHeight`,
+            DEFAULT_INVENTORY_PORTAL_HEIGHT
+        );
+    }
+}
